// webapp/src/app/model/theme.interface.ts
export interface ThemeInterface
{
  name?: string;
  backgroundColor: string;
  petInfoBackgroundColor: string;
  speechBubbleBackgroundColor: string;
  textColor: string;
  primaryColor: string;
  textOnPrimaryColor: string;
  tabBarBackgroundColor: string;
  linkAndButtonColor: string;
  buttonTextColor: string;
  dialogLinkColor: string;
  warningColor: string;
  gainColor: string;
  bonusAndSpiceColor: string;
  bonusAndSpiceSelectedColor: string;
  inputBackgroundColor: string;
  inputTextColor: string;
}